services/svc-fusion/app/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.6.1
httpx==0.28.1
tenacity==9.0.0
python-json-logger==2.0.7
python-jose[cryptography]==3.3.0
azure-storage-blob==12.19.0
